+++ tests/testthat/test_XcmsExperiment-plotting.R
@@ -4,6 +4,7 @@ df <- data.frame(mzML_file = basename(fls),
                  dataOrigin = fls,
                  sample = c("ko15", "ko16", "ko18"))
 mse <- readMsExperiment(spectraFiles = fls, sampleData = df)
+mse_ms2 <- readMsExperiment(msdata::proteomics(full.names=TRUE)[4])
 p <- CentWaveParam(noise = 10000, snthresh = 40, prefilter = c(3, 10000))
 xmse <- findChromPeaks(mse, param = p)
 pdp <- PeakDensityParam(sampleGroups = rep(1, 3))
@@ -50,4 +51,9 @@ test_that("plot,XcmsExperiment and .xmse_plot_xic works", {
 
     tmp <- filterMz(filterRt(xmse, rt = c(2550, 2800)), mz = c(342.5, 344.5))
     plot(tmp)
+})
+
+test_that(".xmse_plot_xic works with ms2 data", {
+  tmp <-  filterMz(filterRt(mse_ms2, rt= c(2160, 2190)), mz = c(990,1000))
+  plot(tmp)
 })