+++ NEWS.md
@@ -1,5 +1,17 @@
 # xcms 4.1
 
+## Changes in version 4.1.13
+
+- Fix for issue #734. XIC plot is is now working with MS2 Data.
+  
+
+## Changes in version 4.1.13
+
+- Add parameter `rtimeDifferenceThreshold` to `ObiwarpParam` allowing to
+  customize the threshold used by obiwarp to determine whether *gaps* are
+  present in the sequence of retention times of a sample. This addresses/fixes
+  issue #739.
+
 ## Changes in version 4.1.12
 
 - Implementation of the `LamaParama` class and method for the `adjustRtime()`